Reduce memory consumption
We don't need to keep residuals for each point and JT if
we calculate `JT * Residuals` in a dedicated value (`J_T_R`)
and update it on each point.
It takes just 5 or 9 floats instead of 5xN (or 9xN) for `JT`
and N for `Residuals`, where N is points count.